# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Adify API",
    "version": "1.0.0",
    "description": "API documentation for Adify - Contextual ad and service engine for parking sessions",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Adify Support",
      "url": "https://adify.example.com",
      "email": "support@adify.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Ad": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The ad ID",
            "example": 1
          },
          "advertiserId": {
            "type": "integer",
            "description": "The advertiser ID",
            "example": 1
          },
          "adUrl": {
            "type": "string",
            "description": "The ad URL",
            "example": "https://example.com/ad1"
          },
          "imageUrl": {
            "type": "string",
            "description": "The ad image URL",
            "example": "https://example.com/image1.jpg"
          },
          "locationIds": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Location IDs where ad is displayed",
            "example": [
              1,
              2,
              3
            ]
          },
          "costToClick": {
            "type": "number",
            "format": "float",
            "description": "Cost per click",
            "example": 0.5
          },
          "categoryId": {
            "type": "integer",
            "description": "Category ID of the ad",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "error": {
            "type": "string",
            "example": "Error details"
          }
        }
      }
    }
  },
  "paths": {
...
    "/recommendations/member/{memberId}": {
      "get": {
        "summary": "Get personalized ad recommendations for a member",
        "tags": [
          "Recommendations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the member"
          },
          {
            "in": "query",
            "name": "locationId",
            "schema": {
              "type": "integer"
            },
            "description": "ID of the location (optional if member has transactions)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Maximum number of ads to return"
          }
        ],
        "responses": {
          "200": {
            "description": "List of recommended ads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ad"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Location ID required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Location ID is required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Member not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Member not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}